Ensure boot handles broken configs better
Previously, booting up with broken vinits would cause nothing to start. We would have a shell, but nothing else. This seems wrong- in fact we want all services that _can_ start to start, and to be able to see in the logs what was broken.

This change will store a field in a Service signifying what the error was, and will move that service to an empty group.

This will allow us to limp a system up as best we can, allowing logins and stuff